Separate common tasks into a base view controller.
